requirements.txt
# All libraries used that need to be installed
 
pynput
requests
Pillow
